consolodate mock tokenizers
